Fix size of allocated dsm
git-svn-id: https://svn.cscs.ch/csviz/trunk/Projects/ICARUS@1618 a8ef729d-7618-0410-a8c4-833010e87eec

+++ vtkDSMManager.cxx
@@ -273,7 +273,7 @@ bool vtkDSMManager::CreateDSM()
   // this->DSMBuffer->DebugOn();
   this->DSMBuffer->SetServiceThreadUseCopy(0);
   // Uniform Dsm : every node has a buffer the same size. (Addresses are sequential)
-  this->DSMBuffer->ConfigureUniform(this->DSMComm, this->GetLocalBufferSizeMBytes()*1024*1024);
+  this->DSMBuffer->ConfigureUniform(this->DSMComm, ((long long) this->GetLocalBufferSizeMBytes())*1024*1024);
   if (this->UpdatePiece == 0) {
     vtkDebugMacro(<< "Length set: " << this->DSMBuffer->GetLength() <<
        ", totalLength set: " << this->DSMBuffer->GetTotalLength() <<
